Bug Fix 🐛 in TaskVC
- The Bug: a bug was crashing the app when we scroll up or down
really fast. The cell's trailing or leading swipe actions gets triggered
and this causes our pendingTask array to run into an indexpath error
mid scroll.

The Fix: Deployed ScrollView Methods to determine when to enable
the tableView's canEditRowAt method. We can only interact with the
cell's swipe actions when the ScrollView is not currently scrolling.